refactor: bilibili-share-to-friends 下沉全部好友面板状态
